lavchik/backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
pydantic
python-dotenv
httpx
psycopg2-binary